fix(selectfield): Selectfiled is not aligned vertically with mdl-textfield
The select field can for some browser and OS combinations, e.g. FireFox and OSX, be off by 2 pixels
compared to the input field. There is no way to fix this without setting an explicit height on both
field types. Since MDL does not set a specific height of the text field, it can not be done for the
select field either. If alignment is required, the user must in his/hers design set a specific
height both for textfields and selectfields.

closes #39