feat(menu-button): Accessible menu button, as defined in in WAI-ARIA Authoring Practices
A menu button is a button that opens a menu. It is often styled as a typical push button with a

downward pointing arrow or triangle to hint that activating the button will display a menu. The menu

button has roles, attributes and behaviour as outlined in WAI-ARIA Authoring Practices

closes #73